# internal/app/docs/swagger.yaml
basePath: /
definitions:
  github_com_eren_dev_go_server_internal_shared_pagination.PaginationInfo:
    properties:
      limit:
        description: Límite de registros por página
        example: 10
        type: integer
      skip:
        description: Registros saltados
        example: 0
        type: integer
      total:
        description: Total de registros
        example: 100
        type: integer
      total_pages:
        description: Total de páginas
        example: 10
        type: integer
    type: object
  github_com_eren_dev_go_server_internal_shared_validation.FieldError:
    properties:
      field:
        description: Nombre del campo con error
        example: email
        type: string
      message:
        description: Mensaje de error
        example: must be a valid email
        type: string
    type: object
  github_com_eren_dev_go_server_internal_shared_validation.ValidationError:
    properties:
      errors:
        description: Lista de errores por campo
        items:
          $ref: '#/definitions/github_com_eren_dev_go_server_internal_shared_validation.FieldError'
        type: array
    type: object
...
  internal_modules_tenant.CreateTenantDTO:
    properties:
      address:
        example: 'Calle 123 #45-67'
        type: string
      commercial_name:
        example: Vet Vida
        type: string
      country:
        example: Colombia
        type: string
      currency:
        example: COP
        type: string
      domain:
        description: Técnico
        example: vetvida
        type: string
      email:
        description: Contacto
        example: contacto@vetvida.com
        type: string
      identification_number:
        example: 900123456-7
        type: string
      industry:
        example: veterinary
        type: string
      logo:
        example: https://example.com/logo.png
        type: string
      name:
        description: Información Básica
        example: Clínica Vet Vida
        minLength: 2
        type: string
      owner_id:
        description: ID del propietario (referencia a users)
        example: 507f1f77bcf86cd799439011
        type: string
      phone:
        example: +57 300 123 4567
        type: string
      secondary_phone:
        example: +57 300 987 6543
        type: string
      timezone:
        example: America/Bogota
        type: string
    required:
    - address
    - commercial_name
    - country
    - currency
    - domain
    - email
    - identification_number
    - industry
    - name
    - owner_id
    - phone
    - timezone
    type: object
  internal_modules_tenant.TenantResponse:
    properties:
      address:
        example: 'Calle 123 #45-67'
        type: string
      commercial_name:
        example: Vet Vida
        type: string
      country:
        example: Colombia
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      currency:
        example: COP
        type: string
      domain:
        example: vetvida
        type: string
      email:
        example: contacto@vetvida.com
        type: string
      id:
        example: 507f1f77bcf86cd799439011
        type: string
      identification_number:
        example: 900123456-7
        type: string
      industry:
        example: veterinary
        type: string
      logo:
        example: https://example.com/logo.png
        type: string
      name:
        example: Clínica Vet Vida
        type: string
      owner_id:
        example: 507f1f77bcf86cd799439011
        type: string
      phone:
        example: +57 300 123 4567
        type: string
      secondary_phone:
        example: +57 300 987 6543
        type: string
      status:
        allOf:
        - $ref: '#/definitions/internal_modules_tenant.TenantStatus'
        example: active
      subscription:
        $ref: '#/definitions/internal_modules_tenant.TenantSubscriptionResponse'
      timezone:
        example: America/Bogota
        type: string
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      usage:
        $ref: '#/definitions/internal_modules_tenant.TenantUsageResponse'
    type: object
  internal_modules_tenant.TenantStatus:
    enum:
    - active
    - inactive
    - trial
    - suspended
    type: string
    x-enum-varnames:
    - Active
    - Inactive
    - Trial
    - Suspended
  internal_modules_tenant.TenantSubscriptionResponse:
    properties:
      billing_status:
        type: string
      external_subscription_id:
        type: string
      mrr:
        type: number
      payment_provider:
        type: string
      plan_id:
        type: string
      subscription_ends_at:
        type: string
      trial_ends_at:
        type: string
    type: object
  internal_modules_tenant.TenantUsageResponse:
    properties:
      last_reset_date:
        type: string
      storage_limit_mb:
        type: integer
      storage_used_mb:
        type: integer
      users_count:
        type: integer
      users_limit:
        type: integer
    type: object
  internal_modules_tenant.UpdateTenantDTO:
    properties:
      address:
        example: 'Calle 123 #45-67'
        type: string
      commercial_name:
        example: Vet Vida
        type: string
      country:
        example: Colombia
        type: string
      currency:
        example: COP
        type: string
      email:
        example: contacto@vetvida.com
        type: string
      identification_number:
        example: 900123456-7
        type: string
      industry:
        example: veterinary
        type: string
      logo:
        example: https://example.com/logo.png
        type: string
      name:
        example: Clínica Vet Vida
        type: string
      phone:
        example: +57 300 123 4567
        type: string
      secondary_phone:
        example: +57 300 987 6543
        type: string
      timezone:
        example: America/Bogota
        type: string
    type: object
  internal_modules_users.CreateUserDTO:
    properties:
      email:
        description: Email del usuario
        example: john@example.com
        type: string
      name:
        description: Nombre del usuario
        example: John Doe
        minLength: 2
        type: string
      password:
        description: Contraseña (mínimo 6 caracteres)
        example: secret123
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  internal_modules_users.PaginatedUsersResponse:
    properties:
      data:
        description: Lista de usuarios
        items:
          $ref: '#/definitions/internal_modules_users.UserResponse'
        type: array
      pagination:
        allOf:
        - $ref: '#/definitions/github_com_eren_dev_go_server_internal_shared_pagination.PaginationInfo'
        description: Información de paginación
    type: object
  internal_modules_users.UpdateUserDTO:
    properties:
      email:
        description: Email del usuario
        example: jane@example.com
        type: string
      name:
        description: Nombre del usuario
        example: Jane Doe
        type: string
    type: object
  internal_modules_users.UserResponse:
    properties:
      created_at:
        description: Fecha de creación
        example: "2024-01-01T00:00:00Z"
        type: string
      email:
        description: Email del usuario
        example: john@example.com
        type: string
      id:
        description: ID único del usuario
        example: 507f1f77bcf86cd799439011
        type: string
      name:
        description: Nombre del usuario
        example: John Doe
        type: string
      updated_at:
        description: Fecha de actualización
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@vetsify.com
    name: Vetsify Support
  description: |-
    API para plataforma veterinaria multi-tenant. Soporta panel de administración (staff) y app móvil (owners).

    ## Autenticación
    Todas las rutas protegidas requieren el header `Authorization: Bearer <token>`.
    - **Staff**: Obtén tokens via `POST /api/auth/login`
    - **Owners (mobile)**: Obtén tokens via `POST /mobile/auth/login`

    ## Multi-tenancy
    Las rutas que operan sobre datos de un tenant específico (patients, species, etc.) requieren el header `X-Tenant-ID` con el ObjectID del tenant.
    Este header debe enviarse en **todas** las peticiones a rutas marcadas con el parámetro `X-Tenant-ID`.
    El frontend puede configurarlo una vez en el interceptor HTTP (axios/fetch).

    ## Roles (RBAC)
    Las rutas del panel admin (`/api/*`) están protegidas por RBAC.
    Roles disponibles: `admin`, `veterinarian`, `receptionist`, `assistant`, `accountant`.

    ## Paginación
    Las rutas que retornan listas soportan `skip` y `limit` como query params.
    Respuesta incluye `pagination: { skip, limit, total, total_pages }`.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Vetsify API
  version: "1.0"
paths:
...
  /api/tenants:
    get:
      consumes:
      - application/json
      description: Obtiene una lista de tenants
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_modules_tenant.TenantResponse'
            type: array
      summary: Listar tenants
      tags:
      - tenant
    post:
      consumes:
      - application/json
      description: Crea un nuevo tenant en el sistema
      parameters:
      - description: Datos del tenant
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_tenant.CreateTenantDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_tenant.TenantResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eren_dev_go_server_internal_shared_validation.ValidationError'
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Crear tenant
      tags:
      - tenant
  /api/tenants/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina un tenant por su ID (soft delete)
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Eliminar tenant
      tags:
      - tenant
    get:
      consumes:
      - application/json
      description: Obtiene un tenant por su ID
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_tenant.TenantResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Obtener tenant
      tags:
      - tenant
    patch:
      consumes:
      - application/json
      description: Actualiza un tenant existente
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      - description: Datos a actualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_tenant.UpdateTenantDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_tenant.TenantResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eren_dev_go_server_internal_shared_validation.ValidationError'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Actualizar tenant
      tags:
      - tenant
...
  /api/users:
    get:
      consumes:
      - application/json
      description: Obtiene una lista paginada de usuarios
      parameters:
      - default: 0
        description: ' '
        in: query
        name: skip
        type: integer
      - default: 10
        description: ' '
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_users.PaginatedUsersResponse'
        "401":
          description: No autorizado
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Listar usuarios
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Crea un nuevo usuario en el sistema
      parameters:
      - description: Datos del usuario
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_users.CreateUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_users.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eren_dev_go_server_internal_shared_validation.ValidationError'
        "401":
          description: No autorizado
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Email ya existe
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Crear usuario
      tags:
      - users
  /api/users/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina un usuario por su ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "401":
          description: No autorizado
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: No encontrado
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Eliminar usuario
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Obtiene un usuario por su ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_users.UserResponse'
        "401":
          description: No autorizado
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: No encontrado
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Obtener usuario
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Actualiza un usuario existente
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Datos a actualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_users.UpdateUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_users.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eren_dev_go_server_internal_shared_validation.ValidationError'
        "401":
          description: No autorizado
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: No encontrado
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Actualizar usuario
      tags:
      - users
  /api/webhooks/{provider}:
    post:
      consumes:
      - application/json
      description: Endpoint para recibir webhooks de proveedores de pago
      parameters:
      - description: Provider name
        enum:
        - wompi
        - stripe
        in: path
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Procesar webhook de pago
      tags:
      - webhooks
...
securityDefinitions:
  Bearer:
    description: 'Ingresa el token con el prefijo Bearer: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"